+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v21.4.12-1.21.4] - 2025-03-11
+### Added
+- Add `ClientAbstractions::copyBakedQuad`
 ### Fixed
 - Fix rare start-up crash when registering generated config translations on NeoForge